requirements.txt
cycler==0.10.0
joblib==0.14.1
kiwisolver==1.1.0
lightgbm==2.3.1
llvmlite==0.31.0
matplotlib==3.2.1
numba==0.48.0
numpy==1.18.2
pandas==1.0.3
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.0
six==1.14.0
umap-learn==0.3.10
